integrate fsdp kernel
