+++ verl/trainer/config/ppo_trainer.yaml
@@ -32,6 +32,8 @@ actor_rollout_ref:
     use_remove_padding: False
     use_liger: False
     use_fused_kernels: False
+    fused_kernel_options:
+      impl_backend: torch # triton, torch
     trust_remote_code: False
   actor:
     strategy: fsdp  # [fsdp, fsdp2], This is for backward-compatibility
@@ -192,6 +194,8 @@ reward_model:
     external_lib: ${actor_rollout_ref.model.external_lib}
     use_remove_padding: False
     use_fused_kernels: ${actor_rollout_ref.model.use_fused_kernels}
+    fused_kernel_options:
+      impl_backend: ${actor_rollout_ref.model.impl_backend} # triton, torch
     trust_remote_code: False
     fsdp_config:
       wrap_policy:
